+++ search-films/src/components/index.jsx
@@ -1,10 +1,10 @@
 import Flex from './Flex';
+import Input from './Input';
 import LinkedMenuItem from './LinkedMenuItem';
 import Logo from './Logo';
 import Menu from './Menu';
 import MovieCard from './MovieCard';
 import Paragraph from './Paragraph';
-import Search from './Search';
 import Title from './Title';
-export { Flex, LinkedMenuItem, Logo, Menu, MovieCard, Paragraph, Search, Title };
+export { Flex, Input, LinkedMenuItem, Logo, Menu, MovieCard, Paragraph, Title };